+++ internal/pkg/models/detail.go
@@ -3,8 +3,8 @@ package models
 import "time"
 
 type Detail struct {
-	ID          uint64    `json:"id"`
+	ID          int64     `json:"id"`
 	Name        string    `json:"name"`
-	Price       float32   `json:"price"`
+	Price       float64   `json:"price"`
 	CreatedTime time.Time `json:"created_time"`
 }
